+++ lib/kcu.rb
@@ -13,7 +13,9 @@ require "kcu/summary"
 require "kcu/services/get_resource_namespace_and_name"
 require "kcu/services/secrets/get_json"
 require "kcu/services/secrets/decode_data"
+require "kcu/services/secrets/get_entry_value"
 require "kcu/actions/list_secret_action"
+require "kcu/actions/get_secret_action"
 require "kcu/client"
 
 Kcu.new.run if $0 == __FILE__