+++ lib/kcu/actions/get_secret_action.rb
@@ -1,0 +1,17 @@
+module Kcu
+  class GetSecretAction
+
+    extend LightService::Organizer
+
+    def self.call(resource, entry_name)
+      ctx = with(resource: resource, entry_name: entry_name).reduce(
+        GetResourceNamespaceAndName,
+        Secrets::GetJson,
+        Secrets::DecodeData,
+        Secrets::GetEntryValue,
+      )
+      ap ctx.entry_value
+    end
+
+  end
+end